requirements.txt
antlr4-python3-runtime==4.9.3
APScheduler==3.10.1
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-pandas==0.6.6
filelock==3.11.0
fonttools==4.39.3
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
#psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
sympy==1.11.1
#torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
